+++ benches/benches/bevy_ecs/components/archetype_updates.rs
@@ -1,4 +1,4 @@
-use bevy_ecs::{component::Component, schedule_v3::Schedule, world::World};
+use bevy_ecs::{component::Component, schedule::Schedule, world::World};
 use criterion::{BenchmarkId, Criterion};
 
 #[derive(Component)]
