+++ crates/bevy_ecs/src/lib.rs
@@ -14,7 +14,7 @@ pub mod query;
 #[cfg(feature = "bevy_reflect")]
 pub mod reflect;
 pub mod removal_detection;
-pub mod schedule_v3;
+pub mod schedule;
 pub mod storage;
 pub mod system;
 pub mod world;
@@ -35,7 +35,7 @@ pub mod prelude {
         event::{Event, EventReader, EventWriter, Events},
         query::{Added, AnyOf, ChangeTrackers, Changed, Or, QueryState, With, Without},
         removal_detection::RemovedComponents,
-        schedule_v3::{
+        schedule::{
             apply_state_transition, apply_system_buffers, common_conditions::*, IntoSystemConfig,
             IntoSystemConfigs, IntoSystemSet, IntoSystemSetConfig, IntoSystemSetConfigs, NextState,
             OnEnter, OnExit, OnUpdate, Schedule, Schedules, State, States, SystemSet,
