+++ crates/bevy_ecs/src/schedule/mod.rs
@@ -2,6 +2,7 @@ mod condition;
 mod config;
 mod executor;
 mod graph_utils;
+#[allow(clippy::module_inception)]
 mod schedule;
 mod set;
 mod state;
@@ -20,7 +21,7 @@ mod tests {
     use std::sync::atomic::{AtomicU32, Ordering};
 
     pub use crate as bevy_ecs;
-    pub use crate::schedule_v3::{IntoSystemConfig, IntoSystemSetConfig, Schedule, SystemSet};
+    pub use crate::schedule::{IntoSystemConfig, IntoSystemSetConfig, Schedule, SystemSet};
     pub use crate::system::{Res, ResMut};
     pub use crate::{prelude::World, system::Resource};
