+++ crates/bevy_ecs/src/system/system_piping.rs
@@ -146,7 +146,7 @@ impl<SystemA: System, SystemB: System<In = SystemA::Out>> System for PipeSystem<
         self.system_b.set_last_change_tick(last_change_tick);
     }
 
-    fn default_system_sets(&self) -> Vec<Box<dyn crate::schedule_v3::SystemSet>> {
+    fn default_system_sets(&self) -> Vec<Box<dyn crate::schedule::SystemSet>> {
         let mut system_sets = self.system_a.default_system_sets();
         system_sets.extend_from_slice(&self.system_b.default_system_sets());
         system_sets
