+++ crates/bevy_input/src/input.rs
@@ -5,7 +5,7 @@ use std::hash::Hash;
 
 // unused import, but needed for intra doc link to work
 #[allow(unused_imports)]
-use bevy_ecs::schedule_v3::State;
+use bevy_ecs::schedule::State;
 
 /// A "press-able" input of type `T`.
 ///
@@ -22,7 +22,7 @@ use bevy_ecs::schedule_v3::State;
 ///
 /// In case multiple systems are checking for [`Input::just_pressed`] or [`Input::just_released`]
 /// but only one should react, for example in the case of triggering
-/// [`State`](bevy_ecs::schedule_v3::State) change, you should consider clearing the input state, either by:
+/// [`State`](bevy_ecs::schedule::State) change, you should consider clearing the input state, either by:
 ///
 /// * Using [`Input::clear_just_pressed`] or [`Input::clear_just_released`] instead.
 /// * Calling [`Input::clear`] or [`Input::reset`] immediately after the state change.
